[run large tests]
yt:git_commit:970b3b171965f106b4a9358dc94a28df4b5cc830
yt:arcadia-sync:project:yt
FFO-KCUF

git-svn-id: svn+ssh://arcadia.yandex.ru/arc/trunk/arcadia/yt/19_4/yt@5763748 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ server/node/exec_agent/slot_manager.h
@@ -42,7 +42,7 @@ public:
 
     std::optional<i64> GetMemoryLimit() const;
 
-    std::optional<i64> GetCpuLimit() const;
+    std::optional<double> GetCpuLimit() const;
 
     bool ExternalJobMemory() const;
